src/main/java/easy/linkedList/Q0021MergeTwoSortedLists/Solution.java
package easy.linkedList.Q0021MergeTwoSortedLists;

/*

https://leetcode.com/problems/merge-two-sorted-lists/description/

Topics: Linked List, Recursion

Time: O(n)
Space: O(1)

 */

import dataStructures.ListNode;

public class Solution {
    public ListNode mergeTwoLists(ListNode list1, ListNode list2) {
        ListNode dummy = new ListNode();
        ListNode tail = dummy;

        while (list1 != null && list2 != null) {
            if (list1.val < list2.val) {
                tail.next = list1;
                list1 = list1.next;
            } else {
                tail.next = list2;
                list2 = list2.next;
            }
            tail = tail.next;
        }

        if (list1 != null) {
            tail.next = list1;
        } else if (list2 != null) {
            tail.next = list2;
        }

        return dummy.next;
    }
}
